Migrate CalculateComputedMetricRankTablesJob
